requirements.txt
requests==2.32.3
python-dotenv==1.1.0
pandas==2.2.3
openpyxl==3.1.5
networkx==3.4.2
geopy==2.4.1
pathlib
scikit-learn==1.6.1
stargazer==0.0.7
statsmodels==0.14.4